% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesquisa.R
\name{plot.pof_analise}
\alias{plot.pof_analise}
\title{Gráfico de uma análise}
\usage{
\method{plot}{pof_analise}(x, ...)
}
\arguments{
\item{x}{Resultado de \code{\link[=pof_analisar]{pof_analisar()}}.}

\item{...}{Não usado.}
}
\value{
Objeto \code{ggplot}.
}
\description{
Gráfico de uma análise
}
