% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler.R
\name{pof_edicao}
\alias{pof_edicao}
\title{Rótulo da edição}
\usage{
pof_edicao(ano)
}
\arguments{
\item{ano}{Ano inicial da edição.}
}
\value{
Texto como \code{"2017-2018"}.
}
\description{
Rótulo da edição
}
\examples{
pof_edicao(2017)
}
