% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desigualdade.R
\name{pof_gini}
\alias{pof_gini}
\title{Coeficiente de Gini ponderado}
\usage{
pof_gini(x, w)
}
\arguments{
\item{x}{Valores (ex. consumo per capita).}

\item{w}{Pesos (ex. peso da UC vezes moradores).}
}
\value{
Gini entre 0 e 1.
}
\description{
Coeficiente de Gini ponderado
}
\examples{
pof_gini(c(1, 1, 1, 1), rep(1, 4))   # 0
pof_gini(c(0, 0, 0, 1), rep(1, 4))   # 0,75
}
