% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desigualdade.R
\name{pof_quantil}
\alias{pof_quantil}
\title{Quantil ponderado}
\usage{
pof_quantil(x, w, p)
}
\arguments{
\item{x}{Valores.}

\item{w}{Pesos.}

\item{p}{Probabilidades.}
}
\value{
Quantis.
}
\description{
Quantil ponderado
}
\examples{
pof_quantil(1:10, rep(1, 10), c(0.1, 0.5, 0.9))
}
