% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler.R
\name{pof_somar_grupos}
\alias{pof_somar_grupos}
\title{Soma os Níveis 1 em grandes grupos de consumo}
\usage{
pof_somar_grupos(b)
}
\arguments{
\item{b}{Base de \code{\link[=pof_ler_edicao]{pof_ler_edicao()}}.}
}
\value{
A base, com as colunas de grupo.
}
\description{
Acrescenta à base uma coluna por grupo de \code{\link[=pof_grupos]{pof_grupos()}} (Alimentação,
Habitação etc.).
}
