% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pofanalise-package.R
\docType{package}
\name{pofanalise-package}
\alias{pofanalise}
\alias{pofanalise-package}
\title{pofanalise: análises da POF harmonizada}
\description{
Funções para analisar as cinco edições da Pesquisa de Orçamentos
Familiares (1987-1988 a 2017-2018) harmonizadas por Arthur Welle.
Comece por \code{\link[=pof_harmonizacao]{pof_harmonizacao()}} e \code{\link[=pof_ler_edicao]{pof_ler_edicao()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/talesalonso1996-ops/pofanalise}
  \item Report bugs at \url{https://github.com/talesalonso1996-ops/pofanalise/issues}
}

}
\author{
\strong{Maintainer}: Tales Alonso \email{tales.alonso1996@gmail.com}

Authors:
\itemize{
  \item Tales Alonso \email{tales.alonso1996@gmail.com}
}

Other contributors:
\itemize{
  \item Arthur Welle (Harmonização dos produtos da POF (Harmoniza_Produtos)) [contributor]
}

}
\keyword{internal}
